# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Post Service API",
    "description": "MSA 아키텍처에서 게시글 관리만을 담당하는 독립적인 서비스 API입니다.",
    "version": "1.0.0",
    "contact": {
      "name": "Post Service Team",
      "email": "post-service@example.com"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
...
  "paths": {
    "/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "게시글 목록 조회",
        "description": "페이지네이션, 검색, 필터링을 지원하는 게시글 목록을 조회합니다.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 1,
            "minimum": 1,
            "description": "페이지 번호"
          },
          {
            "name": "per_page",
            "in": "query",
            "type": "integer",
            "default": 10,
            "minimum": 1,
            "maximum": 50,
            "description": "페이지당 항목 수"
          },
          {
            "name": "q",
            "in": "query",
            "type": "string",
            "description": "검색어 (제목, 내용)"
          },
          {
            "name": "visibility",
            "in": "query",
            "type": "string",
            "enum": ["PUBLIC", "PRIVATE", "UNLISTED"],
            "default": "PUBLIC",
            "description": "게시글 공개 설정"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string",
            "enum": ["PUBLISHED", "DRAFT", "DELETED"],
            "default": "PUBLISHED",
            "description": "게시글 상태"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 목록 조회 성공",
            "schema": {
              "$ref": "#/definitions/PostListResponse"
            }
          },
          "400": {
            "description": "잘못된 요청",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "게시글 생성",
        "description": "새로운 게시글을 생성합니다.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostCreateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "게시글 생성 성공",
            "schema": {
              "$ref": "#/definitions/PostCreateResponse"
            }
          },
          "400": {
            "description": "잘못된 요청 데이터",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/posts/{post_id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "게시글 단건 조회",
        "description": "특정 게시글의 상세 정보를 조회합니다. 조회수는 자동으로 증가합니다.",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "게시글 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 조회 성공",
            "schema": {
              "$ref": "#/definitions/PostResponse"
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없음",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "tags": ["Posts"],
        "summary": "게시글 전체 수정",
        "description": "게시글의 모든 정보를 새로운 데이터로 교체합니다.",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "게시글 ID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 수정 성공",
            "schema": {
              "$ref": "#/definitions/PostUpdateResponse"
            }
          },
          "400": {
            "description": "잘못된 요청 데이터",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없음",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "patch": {
        "tags": ["Posts"],
        "summary": "게시글 부분 수정",
        "description": "게시글의 일부 정보만 수정합니다.",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "게시글 ID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostPartialUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 수정 성공",
            "schema": {
              "$ref": "#/definitions/PostUpdateResponse"
            }
          },
          "400": {
            "description": "잘못된 요청 데이터",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없음",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "게시글 삭제",
        "description": "게시글을 삭제합니다. (소프트 삭제: status를 'DELETED'로 변경)",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "게시글 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 삭제 성공",
            "schema": {
              "$ref": "#/definitions/PostDeleteResponse"
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없음",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/tags": {
      "get": {
        "tags": ["Tags"],
        "summary": "태그 목록 조회",
        "description": "모든 태그 목록을 조회합니다.",
        "responses": {
          "200": {
            "description": "태그 목록 조회 성공",
            "schema": {
              "$ref": "#/definitions/TagListResponse"
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": ["Tags"],
        "summary": "새 태그 생성",
        "description": "새로운 태그를 생성합니다.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TagCreateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "태그 생성 성공",
            "schema": {
              "$ref": "#/definitions/TagCreateResponse"
            }
          },
          "400": {
            "description": "잘못된 요청 데이터",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "PostCreateRequest": {
      "type": "object",
      "required": ["title", "author_id"],
      "properties": {
        "title": {
          "type": "string",
          "description": "게시글 제목",
          "maxLength": 200
        },
        "content_md": {
          "type": "string",
          "description": "마크다운 내용"
        },
        "content_s3url": {
          "type": "string",
          "description": "S3 URL"
        },
        "author_id": {
          "type": "string",
          "description": "작성자 ID"
        },
        "visibility": {
          "type": "string",
          "enum": ["PUBLIC", "PRIVATE", "UNLISTED"],
          "default": "PUBLIC"
        },
        "status": {
          "type": "string",
          "enum": ["PUBLISHED", "DRAFT"],
          "default": "DRAFT"
        }
      }
    },
...
    "TagCreateRequest": {
      "type": "object",
      "required": ["name"],
      "properties": {
        "name": {
          "type": "string",
          "description": "태그 이름",
          "maxLength": 50
        }
      }
    },
...
    "PostListResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "게시글 목록을 성공적으로 조회했습니다"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Post"
          }
        },
        "meta": {
          "type": "object",
          "properties": {
            "page": {
              "type": "integer",
              "description": "현재 페이지"
            },
            "per_page": {
              "type": "integer",
              "description": "페이지당 항목 수"
            },
            "total": {
              "type": "integer",
              "description": "전체 항목 수"
            },
            "pages": {
              "type": "integer",
              "description": "전체 페이지 수"
            }
          }
        }
      }
    },
    "PostResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "게시글을 성공적으로 조회했습니다"
        },
        "data": {
          "$ref": "#/definitions/Post"
        }
      }
    },
    "PostCreateResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "게시글이 성공적으로 생성되었습니다"
        },
        "data": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "생성된 게시글 ID"
            }
          }
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "description": "에러 메시지"
        },
        "error": {
          "type": "object",
          "properties": {
            "code": {
              "type": "integer",
              "description": "HTTP 상태 코드"
            },
            "details": {
              "type": "string",
              "description": "에러 상세 정보"
            }
          }
        }
      }
    }
  }
}